Don't use HTTP::Request::Common directly
Things work fine as is unless you patch this module to use a _content_sub
like cPanel does. Then you get tears due to Common not actually
supporting the passing of subroutines in as Content.

The solution? Use HTTP::Request->new instead, as that works. No idea why
this was not being done the entire time, but eh.

Also reduces some unnecessary code in file_metadata