Subsequent fixes to tests past previous two commits
Interestingly, it appears that *no files can be found* in your google
drive other than the ones you upload. As such, I altered the assertions
at the end which failed to instead look for no entries, and instead look
for content before deleting the file. Let's hope that's "expected
behavior".