Support media upload and multipart upload:
Uploading files is trickier in v3 and I originally got this wrong.

There are three upload types:

1. media upload: Content only, no metadata
2. multipart upload: Content + metadata
3. resumable upload: Much more complicated and left for later

upload_media_file():

The media upload now available as "upload_media_file()" just uploads
the file using a POST request. There's no metadata other than the
mimetype, which we serve as extra headers. We also set the size
correctly.

Since we're using the "_content_sub()" method, we support progressive
reading of the file, which is great.

upload_multipart_file():

The multipart upload is trickier: We support body parameters as well
and need to report the mime types (one mime type for JSON body
parameters, one mime type for the file, and one mime type for the
multipart/realted envelope).

In this case, we have to read the entire file because the
"_content_sub()" method will not work for it. Wer will need to wrap
the entire envelope creation with it to use it and that will require
a  different method that I haven't written yet.

upload_resumable_file():

This is a much more complicated upload type that requires a lot more
work and is not implemented yet.